# Make sure no AST types are bare slices