--- a/go/vt/schemadiff/schema.go
+++ b/go/vt/schemadiff/schema.go
@@ -1127,7 +1127,7 @@ func (s *Schema) getTableColumnNames(t *CreateTableEntity) (columnNames []*sqlpa
 // getViewColumnNames returns the names of aliased columns returned by a given view.
 func (s *Schema) getViewColumnNames(v *CreateViewEntity, schemaInformation *declarativeSchemaInformation) ([]*sqlparser.IdentifierCI, error) {
 	var columnNames []*sqlparser.IdentifierCI
-	for _, node := range v.Select.GetColumns() {
+	for _, node := range v.Select.GetColumns().Exprs {
 		switch node := node.(type) {
 		case *sqlparser.StarExpr:
 			if tableName := node.TableName.Name.String(); tableName != "" {