… go/test/endtoend/vtgate/queries/random/query_gen.go
… go/tools/asthelpergen/asthelpergen_test.go
… go/tools/asthelpergen/integration/integration_rewriter_test.go
… go/tools/asthelpergen/rewrite_gen.go
… go/vt/sqlparser/analyzer_test.go
… go/vt/sqlparser/ast_copy_on_rewrite.go
… go/vt/sqlparser/ast_equals.go
… go/vt/sqlparser/ast_format_fast.go
… go/vt/sqlparser/ast_format_test.go
… go/vt/sqlparser/ast_funcs.go
… go/vt/sqlparser/ast_rewrite.go
… go/vt/sqlparser/ast_visit.go
… go/vt/sqlparser/cached_size.go
… go/vt/sqlparser/normalizer.go
… go/vt/sqlparser/parsed_query_test.go
… go/vt/sqlparser/parser.go
… go/vt/sqlparser/precedence_test.go
… go/vt/sqlparser/predicate_rewriting.go
… go/vt/sqlparser/random_expr.go
… go/vt/sqlparser/rewriter_api.go
… go/vt/sqlparser/rewriter_test.go
… go/vt/sqlparser/sql.go
… go/vt/sqlparser/sql.y
… go/vt/sqlparser/tracked_buffer.go
… go/vt/vtctl/workflow/materializer.go
… go/vt/vtctl/workflow/stream_migrator.go
… go/vt/vtctl/workflow/utils.go
… go/vt/vtexplain/vtexplain_vttablet.go
… go/vt/vtgate/engine/insert_test.go
… go/vt/vtgate/evalengine/expr_env_test.go