## Description
Having types be bare slices, such as `sqlparser.Exprs` and `sqlparser.SelectExprs`, cause confusion with the generated rewriters, and doesn't really improve anything, so talking it over with @vmg we thought we should make sure that all AST nodes are structs and not slices.

Running benchmarks show some nice numbers:
```
                                 │ ../main.bench │       ../less-slices.bench        │
                                 │    sec/op     │   sec/op     vs base              │
OLTP-8                               223.5µ ± 0%   217.6µ ± 2%  -2.62% (p=0.002 n=6)
TPCC-8                               1.596m ± 2%   1.566m ± 1%  -1.89% (p=0.002 n=6)
TPCH-8                               9.678m ± 4%   9.589m ± 0%  -0.92% (p=0.002 n=6)
Planner/from_cases.json-gen4-8       6.925m ± 1%   6.786m ± 1%  -2.00% (p=0.002 n=6)
Planner/filter_cases.json-gen4-8     146.0m ± 0%   145.5m ± 1%       ~ (p=0.310 n=6)
Planner/large_cases.json-gen4-8      444.5µ ± 0%   446.1µ ± 1%  +0.35% (p=0.015 n=6)
Planner/aggr_cases.json-gen4-8       12.45m ± 1%   12.18m ± 2%  -2.15% (p=0.009 n=6)
Planner/select_cases.json-gen4-8    10.013m ± 1%   9.730m ± 1%  -2.83% (p=0.002 n=6)
Planner/union_cases.json-gen4-8      3.564m ± 0%   3.478m ± 0%  -2.42% (p=0.002 n=6)
SemAnalysis-8                        27.94m ± 2%   27.61m ± 0%  -1.20% (p=0.002 n=6)
BaselineVsMirrored/Baseline-8        113.7µ ± 1%   109.4µ ± 0%  -3.81% (p=0.002 n=6)
BaselineVsMirrored/Mirrored-8        312.4µ ± 1%   304.5µ ± 0%  -2.53% (p=0.002 n=6)
geomean                              3.064m        3.007m       -1.87%

                                 │ ../main.bench │        ../less-slices.bench        │
                                 │     B/op      │     B/op      vs base              │
OLTP-8                              151.2Ki ± 0%   149.5Ki ± 0%  -1.18% (p=0.002 n=6)
TPCC-8                              993.6Ki ± 0%   977.7Ki ± 0%  -1.60% (p=0.002 n=6)
TPCH-8                              5.226Mi ± 0%   5.182Mi ± 0%  -0.85% (p=0.002 n=6)
Planner/from_cases.json-gen4-8      4.421Mi ± 0%   4.343Mi ± 0%  -1.76% (p=0.002 n=6)
Planner/filter_cases.json-gen4-8    19.37Mi ± 0%   19.27Mi ± 0%  -0.50% (p=0.002 n=6)
Planner/large_cases.json-gen4-8     273.7Ki ± 0%   273.2Ki ± 0%  -0.19% (p=0.002 n=6)
Planner/aggr_cases.json-gen4-8      7.128Mi ± 0%   6.997Mi ± 0%  -1.84% (p=0.002 n=6)
Planner/select_cases.json-gen4-8    6.142Mi ± 0%   6.009Mi ± 0%  -2.18% (p=0.002 n=6)
Planner/union_cases.json-gen4-8     2.188Mi ± 0%   2.139Mi ± 0%  -2.23% (p=0.002 n=6)
SemAnalysis-8                       18.98Mi ± 0%   18.82Mi ± 0%  -0.87% (p=0.002 n=6)
BaselineVsMirrored/Baseline-8       81.69Ki ± 0%   79.55Ki ± 0%  -2.62% (p=0.002 n=6)
BaselineVsMirrored/Mirrored-8       201.6Ki ± 0%   197.8Ki ± 0%  -1.89% (p=0.002 n=6)
geomean                             1.676Mi        1.651Mi       -1.48%

                                 │ ../main.bench │       ../less-slices.bench        │
                                 │   allocs/op   │  allocs/op   vs base              │
OLTP-8                               3.705k ± 0%   3.637k ± 0%  -1.84% (p=0.002 n=6)
TPCC-8                               23.71k ± 0%   23.08k ± 0%  -2.67% (p=0.002 n=6)
TPCH-8                               131.8k ± 0%   130.0k ± 0%  -1.36% (p=0.002 n=6)
Planner/from_cases.json-gen4-8      102.61k ± 0%   99.55k ± 0%  -2.99% (p=0.002 n=6)
Planner/filter_cases.json-gen4-8     508.9k ± 0%   505.1k ± 0%  -0.75% (p=0.002 n=6)
Planner/large_cases.json-gen4-8      10.61k ± 0%   10.60k ± 0%  -0.14% (p=0.002 n=6)
Planner/aggr_cases.json-gen4-8       160.4k ± 0%   155.1k ± 0%  -3.35% (p=0.002 n=6)
Planner/select_cases.json-gen4-8     141.0k ± 0%   135.7k ± 0%  -3.75% (p=0.002 n=6)
Planner/union_cases.json-gen4-8      50.66k ± 0%   48.70k ± 0%  -3.87% (p=0.002 n=6)
SemAnalysis-8                        231.3k ± 0%   224.9k ± 0%  -2.78% (p=0.002 n=6)
BaselineVsMirrored/Baseline-8        2.017k ± 0%   1.932k ± 0%  -4.21% (p=0.002 n=6)
BaselineVsMirrored/Mirrored-8        5.578k ± 0%   5.434k ± 0%  -2.58% (p=0.002 n=6)
geomean                              40.44k        39.42k       -2.53%
```


## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [x] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [x] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->